// src/app/Models/contacto-alumno.ts
export class ContactoAlumno {
    idContacto: number;
    email: string;
    codigoPostal: string;
    entidadFederativa: string;
    municipio: string;
    colonia: string;
    calle: string;
    numeroExterior: number;
    numeroInterior: number;
    ayudaDirecion: string;
    numeroTelefono: string;
    numeroTelefonoSecundario: string;
    tutor: string;
    telefonoTutor: string;
}